+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # CHANGELOG
 ## [Unreleased]
-- `partiallyEqual()` function to partially compare objects
+
+## [1.1.0](../../compare/1.0.0..1.1.0) - 2021-11-05
+### Added
+- `strictlyEqual()` function
+- `partlyEqual()` function
+
+### Changed
+- `equal()` has an ability to compare objects partially in addition to strict comparison
 
 ## [1.0.0](../../tree/1.0.0) - 2021-11-03
 Release